refactor: 중복로직 merge
