+++ FoodDiary/App/Sources/AppFlowController.swift
@@ -49,28 +49,14 @@ private extension AppFlowController {
     }
     
     func createMainView() -> UIViewController {
-        guard let weeklyCalendarUseCase = try? container.resolve(
-            FetchWeeklyCalendarUseCase<MockFoodRecordRepository>.self
-        ) else {
-            fatalError("FetchWeeklyCalendarUseCase not registered")
-        }
-
         guard let fetchFoodImageAssetUseCase = try? container.resolve(
             FetchFoodImageAssetUseCase<FoodImageAssetFetcher<TFLiteFoodClassifier, UIImageLoader>>.self
         ) else {
             fatalError("FetchFoodImageAssetUseCase not registered")
         }
 
-        guard let fetchFoodRecordsUseCase = try? container.resolve(
-            FetchFoodRecordsUseCase<MockFoodRecordRepository>.self
-        ) else {
-            fatalError("FetchFoodRecordsUseCase not registered")
-        }
-
-        guard let saveFoodRecordUseCase = try? container.resolve(
-            SaveFoodRecordUseCase<MockFoodRecordRepository>.self
-        ) else {
-            fatalError("SaveFoodRecordUseCase not registered")
+        guard let foodRecordRepository = try? container.resolve(MockFoodRecordRepository.self) else {
+            fatalError("MockFoodRecordRepository not registered")
         }
 
         guard let imageProvider = try? container.resolve(UIImageLoader.self) else {
@@ -84,10 +70,8 @@ private extension AppFlowController {
         }
 
         let viewModel = WeeklyCalendarViewModel(
-            fetchWeeklyCalendarUseCase: weeklyCalendarUseCase,
             fetchFoodImageAssetUseCase: fetchFoodImageAssetUseCase,
-            fetchFoodRecordsUseCase: fetchFoodRecordsUseCase,
-            saveFoodRecordUseCase: saveFoodRecordUseCase,
+            foodRecordRepository: foodRecordRepository,
             requestPhotoAuthorizationUseCase: requestPhotoAuthorizationUseCase,
             imageProvider: imageProvider
         )
